// src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: string;
};

export type ClassType = 't' | 'p' | 'l' | 'test';

export type ClassDay = 'lu' | 'ma' | 'mi' | 'ju' | 'vi' | 'sa' | 'do';

export type SubjectSchedule = {
    defined: true;
    type: ClassType;
    day: ClassDay;
    group?: number;
    blocks: Array<number>;
    classroom: string;
};

export type SubjectScheduleTbd = {
    defined: false;
    type?: ClassType;
    group?: number;
};

export type Subject = {
    schedule: Array<SubjectSchedule | SubjectScheduleTbd>;
    code: number;
    name: string;
    credits?: number;
    section: number;
};

export type Schedule = {
    schedules: {
        [key: string]: Subject;
    };
    updating: boolean;
};

export type HttpException = {
    /**
     * Short description of the HTTP error
     */
    message: string;
    /**
     * The HTTP status code
     */
    statusCode: number;
    /**
     * Name of the HTTP status code
     */
    error?: string;
};

export type GetScheduleData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/schedule';
};

export type GetScheduleErrors = {
    /**
     * An error occurred while retrieving the schedule.
     */
    500: HttpException;
};

export type GetScheduleError = GetScheduleErrors[keyof GetScheduleErrors];

export type GetScheduleResponses = {
    /**
     * An object containing the entire schedule.
     */
    200: Schedule;
};

export type GetScheduleResponse = GetScheduleResponses[keyof GetScheduleResponses];